Solved issue #70
This pull request adds code to prevent the user from making or renaming a note with characters forbidden for file names (includes "/"). To allow notes to have "/"s in their names, we would need a new way of storing notes.

Other Changes:
Corrected the end html tag in reader.js